feat(ui): global variables component/container distinction
