+++ nx/libs/feature-variables/src/lib/components/variables-edit/variables-edit.component.html
@@ -1,0 +1,42 @@
+<mat-tab-group animationDuration="0ms"
+               [selectedIndex]="activeGroupNameIndex"
+               (selectedIndexChange)="activeGroupNameIndex = $event"
+               (selectedTabChange)="onTabChange($event)"
+>
+  <mat-tab *ngFor="let groupName of globalVariableGroupsNames; let i=index;"
+           label="{{ groupName }}"
+  >
+    <ng-template *ngIf="groupName.length == 0" mat-tab-label>
+      <mat-form-field appearance="standard">
+        <input matInput
+               id="newGroupNameInput"
+               placeholder="new group"
+               required
+        />
+      </mat-form-field>
+    </ng-template>
+    <ng-template *ngIf="groupName.length != 0" mat-tab-label>
+      <div *ngIf="i != groupNameRenameIndex" (click)="onGroupNameClick(i)">{{ groupName }}</div>
+      <mat-form-field *ngIf="i == groupNameRenameIndex" appearance="standard">
+        <input matInput
+               id="renameGroupNameInput"
+               value="{{groupName}}"
+               required
+        />
+      </mat-form-field>
+    </ng-template>
+  </mat-tab>
+</mat-tab-group>
+
+<div class="my-code-editor" *ngIf="globalVariableGroupsNames.length > 0">
+  <td-code-editor #monaco
+                  fxFlex="grow"
+                  editorStyle="border:0;"
+                  language="json"
+                  [editorOptions]="{ fontSize: 12, automaticLayout: true }"
+                  [theme]="options.theme === 'light' ? 'vs' : 'vs-dark'"
+                  [value]="''"
+                  (editorInitialized)="editorInitialized($event)"
+                  (editorConfigurationChanged)="monacoEditorConfigChanged(monaco.theme)"
+  ></td-code-editor>
+</div>
