+++ nx/libs/feature-variables/src/lib/components/variables-edit/variables-edit.component.scss
@@ -1,0 +1,3 @@
+.my-code-editor {
+  height: calc(100vh - 216px);
+}
