+++ nx/libs/feature-variables/src/lib/containers/variables/variables.component.scss
@@ -1,8 +1,7 @@
-
 .zi-2 {
   z-index: 2;
 }
 
-.my-code-editor {
-  height: calc(100vh - 216px);
+.mr-5 {
+  margin-right: 5px;
 }
